# fix compat mode perf_cpu_hotplug